==> notebooks/02_aggregation.ipynb <==
# Atlas data aggregation

%reset

### Setup

# Load environment variables
    # Dependencies:
    # - python-dotenv

from dotenv import load_dotenv, dotenv_values
import os

# Load env vars (manual)
    # load_dotenv("../.env/master.env")
    # maps_key = os.getenv("maps_key")

# Load all variables from the .env file into a dictionary
env_vars = dotenv_values("../.env/master.env")

# Set them as environment variables
for key, value in env_vars.items():
    os.environ[key] = value
    print(f"Loaded: {key}")

# Print
print(f"\nSuccessfully loaded {len(env_vars)} environment variables from master.env")

# print(env_vars.get("VAR1")) # call variable
# VAR1 = env_vars.get("VAR1")

### Results: Fetch and clean

# Fetch measurement data from RIPE Atlas
    # Dependencies:
    # - RIPE Atlas API

import pandas as pd
import requests
import json, pprint
from ripe.atlas.sagan import PingResult, TracerouteResult, DnsResult, SslResult

api_key = env_vars.get("atlas_key")

# atlasregion = "aws-fra"  # "AWS Frankfurt"
# measurement_id = 155973633 

atlasregion = "aws-sto"  # "AWS Stockholm"
measurement_id = 155991104

# TXT response

import json, pprint
from ripe.atlas.sagan import PingResult, TracerouteResult, DnsResult, SslResult

url = f"https://atlas.ripe.net/api/v2/measurements/{measurement_id}/results/"
params = {
    "key": api_key,
    # optional filters:
    # "start": 1700000000,  # Unix timestamp
    # "stop":  1700003600,
    # "probe_ids": "1,2,3",
    "page_size": 100,
    "format": "txt"  # no pagination necessary
}

response = requests.get(url, params=params)

results = []
for line in response.text.strip().splitlines():
    if line:  # skip empty lines
        results.append(json.loads(line))

# Confirm fetch
print("Length:", len(results))

# Parse & Print raw TXT data per dict element

import json, pprint
from ripe.atlas.sagan import PingResult

parsed_results = [PingResult(r) for r in results]

# pprint.pprint(vars(parsed_results[1]))  # 0,1,2,3,4,5 results in dict

# Preview parsed results

parsed_results = [PingResult(r) for r in results]

for result in parsed_results[:2]:  # preview first 2 results only
    msm_id = result.raw_data.get('msm_id')
    type = result.raw_data.get('type')
    step = result.raw_data.get('step')
    proto = result.raw_data.get('proto')
    prb_id = result.raw_data['prb_id']
    dst_addr = result.raw_data.get('dst_addr')
    dst_name = result.raw_data.get('dst_name')
    af = result.raw_data.get('af')
    size = result.raw_data.get('size')
    avg_rtt = result.raw_data['avg']  # custom name
    min_rtt = result.raw_data.get('min')
    med_rtt = result.rtt_median  # custom name
    max_rtt = result.raw_data.get('max')
    sent = result.raw_data['sent']
    rcvd = result.raw_data['rcvd']
    lost = sent - rcvd  # custom name
    ttl = result.raw_data.get('ttl')

    # print(f"Probe {result.raw_data['prb_id']}\nAvg RTT: {result.raw_data['avg']} ms\nPackets sent: {result.raw_data['sent']}\nLoss: {lost}\n") # directly access dict
    print(f"Probe {prb_id}\nAvg RTT: {avg_rtt} ms\nSent: {sent}\nLost: {lost}\n") # use custom variables, preferred

# Create df_results
# for each result in parsed_results, extract relevant fields and create a DataFrame

import pandas as pd

df_results = pd.DataFrame([
    {
        'msm_id': result.raw_data.get('msm_id'),
        'type': result.raw_data.get('type'),
        'step': result.raw_data.get('step'),
        'proto': result.raw_data.get('proto'),
        'prb_id': result.raw_data['prb_id'],
        'dst_addr': result.raw_data.get('dst_addr'),
        'dst_name': result.raw_data.get('dst_name'),
        'af': result.raw_data.get('af'),
        'size': result.raw_data.get('size'),
        'avg_rtt': result.raw_data['avg'],
        'min_rtt': result.raw_data.get('min'),
        'med_rtt': result.rtt_median,
        'max_rtt': result.raw_data.get('max'),
        'sent': result.raw_data['sent'],
        'rcvd': result.raw_data['rcvd'],
        'lost': result.raw_data['sent'] - result.raw_data['rcvd'],
        'ttl': result.raw_data.get('ttl')
    }
    for result in parsed_results
])

df_results.head()

# Further: Augmet with geo from ../data/processed/probes_de_20260218-061315.csv

### Preprocess other datasets

# Load probes sample

import pandas as pd

df_probes_sample = pd.read_csv(
    '../data/processed/probes_sample.csv',
    dtype={
        'mun_key': str
        }
    )
print(f"Cols: {df_probes_sample.columns.tolist()}\n")
print(f"Types:\n{df_probes_sample.dtypes}\n")
df_probes_sample.head()

# Load municipality sample

import pandas as pd

df_mun_sample = pd.read_csv(
    '../data/processed/municipalities_full_sample.csv',
    dtype={
        'mun_key': str
        }
    )
print(f"Cols: {df_mun_sample.columns.tolist()}\n")
print(f"Types:\n{df_mun_sample.dtypes}\n")
df_mun_sample.head()

# Reshape probe sample to long format

import pandas as pd

df_probes_sample_long = df_probes_sample.melt(id_vars='mun_key', var_name='temp', value_name='value')

df_probes_sample_long[['col', 'num']] = df_probes_sample_long['temp'].str.rsplit('-', n=1, expand=True)

df_probes_sample_long = df_probes_sample_long.drop(columns='temp').pivot_table(
    index=['mun_key', 'num'], columns='col', values='value', aggfunc='first'
).reset_index().drop(columns='num').rename_axis(None, axis=1)

df_probes_sample_long = df_probes_sample_long.rename(columns={'probe': 'probe_id'})
df_probes_sample_long['probe_id'] = df_probes_sample_long['probe_id'].astype('int64')

print(f"Cols: {df_probes_sample_long.columns.tolist()}\n")
print(f"Types:\n{df_probes_sample_long.dtypes}\n")
df_probes_sample_long.head()

# Write to csv

df_probes_sample_long.to_csv(
    '../data/processed/probes_sample_long.csv',
    index=False
)

print(f"Saved df_probes_sample_long to file.")
print(f"Shape: {df_probes_sample_long.shape}")

### Join

# Join df_probes_sample_long with df_results on probe_id
df_merged = df_probes_sample_long.merge(
    df_results, 
    left_on='probe_id', 
    right_on='prb_id', 
    how='left'
)

# Join with df_mun_sample on mun_key
df_merged = df_merged.merge(
    df_mun_sample, 
    on='mun_key', 
    how='left'
)

# Drop all without probe_id (no match in results)
df_merged = df_merged.dropna(subset=['prb_id'])  # use col name from df "results"

# Rename lat and lon for clarity
df_merged = df_merged.rename(columns={
    'lat': 'mun_lat',
    'lon': 'mun_lon'
})

# Fix dtypes
df_merged = df_merged.astype({  # nullable integers (Int64) for graceful handling of NaN
    'mun_key': 'str',
    'distance': 'float64',
    'probe_id': 'int64',
    'msm_id': 'Int64',
    'type': 'str',
    'step': 'str',
    'proto': 'str',
    'dst_addr': 'str',
    'dst_name': 'str',
    'af': 'Int64',
    'size': 'Int64',
    'avg_rtt': 'float64',
    'min_rtt': 'float64',
    'med_rtt': 'float64',
    'max_rtt': 'float64',
    'sent': 'Int64',
    'rcvd': 'Int64',
    'lost': 'Int64',
    'ttl': 'Int64',
    'sample_type': 'str',
    'mun_name_short': 'str',
    'mun_pop_cen22': 'int64',
    'state': 'str',
    'mun_lat': 'float64',
    'mun_lon': 'float64'
})

print(f"Cols: {df_merged.columns.tolist()}\n")
print(f"Shape: {df_merged.shape}\n")
df_merged.head()


# Write to csv

df_merged.to_csv(
    f'../data/processed/df_merged_{atlasregion}.csv',
    index=False
)

print(f"Saved to file.")
print(f"Shape: {df_merged.shape}")

# Combine ../data/processed/df_merged_aws-sto.csv and df_merged_aws-fra.csv into one df_merged_all
